Refactor monomorphization/type resolving code out of `TypeCheckContext`.
